# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "server API for stock web",
        "title": "Stock-Show Web API",
        "termsOfService": "http://github.com/greetlist/",
        "contact": {
            "name": "Greetlist",
            "url": "http://github.com/greetlist",
            "email": "lrt12250914@outlook.com"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "User Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User Login",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "login",
                        "name": "request_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LoginResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "description": "User Logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User Logout",
                "operationId": "logout",
                "parameters": [
                    {
                        "description": "Logout",
                        "name": "request_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LogoutResponse"
                        }
                    }
                }
            }
        },
...
        "/api/user/addPreferStock": {
            "post": {
                "description": "User AddPreferStock",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User AddPreferStock",
                "operationId": "addPreferStock",
                "parameters": [
                    {
                        "description": "login",
                        "name": "request_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddPreferStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AddPreferStockResponse"
                        }
                    }
                }
            }
        },
        "/api/user/deletePreferStock": {
            "post": {
                "description": "User DeletePreferStock",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User DeletePreferStock",
                "operationId": "deletePreferStock",
                "parameters": [
                    {
                        "description": "login",
                        "name": "request_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeletePreferStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DeletePreferStockResponse"
                        }
                    }
                }
            }
        },
        "/api/user/getPreferStock": {
            "post": {
                "description": "User GetPreferStock",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "User GetPreferStock",
                "operationId": "getPreferStock",
                "parameters": [
                    {
                        "description": "login",
                        "name": "request_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetPreferStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetPreferStockResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AddPreferStockRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "stock_code": {
                    "type": "string"
                }
            }
        },
        "model.AddPreferStockResponse": {
            "type": "object",
            "properties": {
                "error_msg": {
                    "type": "string"
                },
                "is_success": {
                    "type": "boolean"
                }
            }
        },
        "model.DeletePreferStockRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "stock_code": {
                    "type": "string"
                }
            }
        },
        "model.DeletePreferStockResponse": {
            "type": "object",
            "properties": {
                "error_msg": {
                    "type": "string"
                },
                "is_success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.FollowStock": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "exchange": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stock_code": {
                    "type": "string"
                }
            }
        },
...
        "model.GetPreferStockRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                }
            }
        },
        "model.GetPreferStockResponse": {
            "type": "object",
            "properties": {
                "error_msg": {
                    "type": "string"
                },
                "follow_stock_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.FollowStock"
                    }
                },
                "is_success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.LoginRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "passwd": {
                    "type": "string"
                }
            }
        },
        "model.LoginResponse": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                },
                "login_succ": {
                    "type": "boolean"
                }
            }
        },
        "model.LogoutRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                }
            }
        },
        "model.LogoutResponse": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                },
                "logout_succ": {
                    "type": "boolean"
                }
            }
        },
...
    }
}